+++ Sources/KeyboardKit/Haptics/HapticFeedback.swift
@@ -27,6 +27,11 @@ public enum HapticFeedback: CaseIterable, Equatable {
     selectionChanged,
     
     none
+    
+    /**
+     The standard player that is used for haptic feedback.
+     */
+    public static var player: HapticFeedbackPlayer = StandardHapticFeedbackPlayer()
 }
 
 
@@ -39,14 +44,7 @@ public extension HapticFeedback {
     }
     
     static func prepare(_ feedback: HapticFeedback) {
-        switch feedback {
-        case .error, .success, .warning: notificationGenerator.prepare()
-        case .lightImpact: lightImpactGenerator.prepare()
-        case .mediumImpact: mediumImpactGenerator.prepare()
-        case .heavyImpact: heavyImpactGenerator.prepare()
-        case .selectionChanged: selectionGenerator.prepare()
-        case .none: return
-        }
+        player.prepare(feedback)
     }
     
     func trigger() {
@@ -54,37 +52,6 @@ public extension HapticFeedback {
     }
     
     static func trigger(_ feedback: HapticFeedback) {
-        switch feedback {
-        case .error: triggerNotification(.error)
-        case .success: triggerNotification(.success)
-        case .warning: triggerNotification(.warning)
-        case .lightImpact: lightImpactGenerator.impactOccurred()
-        case .mediumImpact: mediumImpactGenerator.impactOccurred()
-        case .heavyImpact: heavyImpactGenerator.impactOccurred()
-        case .selectionChanged: selectionGenerator.selectionChanged()
-        case .none: return
-        }
+        player.play(feedback)
     }
-}
-
-
-// MARK: - Private Trigger Functions
-
-private extension HapticFeedback {
-    
-    static func triggerNotification(_ notification: UINotificationFeedbackGenerator.FeedbackType) {
-        notificationGenerator.notificationOccurred(notification)
-    }
-}
-
-
-// MARK: - Private Generators
-
-private extension HapticFeedback {
-    
-    private static var notificationGenerator = UINotificationFeedbackGenerator()
-    private static var lightImpactGenerator = UIImpactFeedbackGenerator(style: .light)
-    private static var mediumImpactGenerator = UIImpactFeedbackGenerator(style: .medium)
-    private static var heavyImpactGenerator = UIImpactFeedbackGenerator(style: .heavy)
-    private static var selectionGenerator = UISelectionFeedbackGenerator()
 }